Module-3/api/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
scikit-learn==0.24.2
scipy==1.7.1
scikit-image==0.18.3
pandas==1.3.3
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
pydantic==1.8.2
matplotlib==3.4.3
seaborn==0.11.2 